Pass a valid child bci to afterChild for short circuit ops
Resolves a bug where we'd always pass bci-3. For s.c. ops that returned the original (non-converted) value, this was an arbitrary location in the bytecode.

In Python, one instruction attempted to "quicken" its short-circuiting child operation, but the incorrect bci pointed at an instruction's node index immediate. The node index matched a quickenable instruction opcode, and so the node index was overwritten with a different index.

I tried to use "-1" when boolean was not boxing eliminated, but our bytecode validation expects some ops (e.g. local stores) to have valid child bci's, so it was easier to just always store and use the childBci.